Move TextCursor to text_cursor.dart.
