+++ patterns/state/manipulator_state/states/selections/text/text_resize_marker_state.dart
