+++ patterns/state/manipulator_state/states/selections/text_resize_state.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 import '../../shapes/text_shape.dart';
 import '../_/sub_states/parent_state.dart';
 import 'text_change_state.dart';
-import 'text_resize_marker/text_resize_marker_state.dart';
+import 'text/text_resize_marker_state.dart';
 
 class TextResizeState extends ParentState<TextShape> {
   TextResizeState({required super.selectedShape}) {
